requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astor==0.8.1
audioread==2.1.8
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
joblib==0.16.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.2.2
numba==0.51.2
numpy==1.19.2
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
pooch==1.2.0
protobuf==3.13.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
